Use fraggle rock API for conducting lighthouse tests In the legacy lighthouse flow, a browser needs to be started with RDP enabled and lighthouse starts testing by firing CDP commands over this RDP. Lighthouse team has been working on the fraggle rock API with flow support and several changes (see: https://github.com/GoogleChrome/lighthouse/issues/11313). This also includes a navigation API that can connect to existing puppeteer page instance for performing lighthouse tests. Using this navigation API in playwright lighthouse, modified the playwright object to mock puppeteer specific API.
